Added Address Exception Handling

+++ src/main/java/com/cs/service/AddressServiceImp.java
@@ -7,8 +7,18 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.cs.bean.Address;
+import com.cs.exception.AddressNotFoundException;
 import com.cs.repository.IAddressRepository;
 
+/**
+ * This class is used to provide services to AddressController so that we can
+ * add, delete, update and retrieve the address into or from database
+ * 
+ * @author Rohith(Employee id: 46191986)
+ * @version 1.0.0
+ * @since 28-03-2022
+ *
+ */
 @Service
 public class AddressServiceImp implements IAddressService {
 
@@ -25,33 +35,48 @@ public class AddressServiceImp implements IAddressService {
 	@Override
 	public Address getAddresById(int id) {
 
-		return addressRepo.findById(id).get();
+		Optional<Address> address = addressRepo.findById(id);
+		if (address.isEmpty()) {
+			throw new AddressNotFoundException("Address with the given id: " + id + " does not exist.");
+		}
+		return address.get();
 	}
 
 	// retrieves the list of all available addresses in database
 	@Override
 	public List<Address> getAllAddresses() {
-		return addressRepo.findAll();
+		List<Address> addresses = addressRepo.findAll();
+		if (addresses.isEmpty()) {
+			throw new AddressNotFoundException(
+					"The address list is empty in the database. Please add some before retrieving");
+		}
+		return addresses;
 	}
 
 	// updates the existing address of given id in database.
 	@Override
 	public Address updateAddressById(int id, Address address) {
 
-		Optional<Address> opt = addressRepo.findById(id);
-		opt.get().setBuildingName(address.getBuildingName());
-		opt.get().setCity(address.getCity());
-		opt.get().setColony(address.getColony());
-		opt.get().setCountry(address.getCountry());
-		opt.get().setFlatNum(address.getFlatNum());
-		opt.get().setPincode(address.getPincode());
-		return addressRepo.save(opt.get());
+		Optional<Address> oldAddress = addressRepo.findById(id);
+		if (oldAddress.isEmpty()) {
+			throw new AddressNotFoundException("Address with the given id: " + id + " does not exist.");
+		}
+		oldAddress.get().setBuildingName(address.getBuildingName());
+		oldAddress.get().setCity(address.getCity());
+		oldAddress.get().setColony(address.getColony());
+		oldAddress.get().setCountry(address.getCountry());
+		oldAddress.get().setFlatNum(address.getFlatNum());
+		oldAddress.get().setPincode(address.getPincode());
+		return addressRepo.save(oldAddress.get());
 	}
 
 	// deletes address of given id from database.
 	@Override
 	public Address deleteAddressById(int id) {
 		Optional<Address> address = addressRepo.findById(id);
+		if (address.isEmpty()) {
+			throw new AddressNotFoundException("Address with the given id: " + id + " does not exist.");
+		}
 		addressRepo.deleteById(id);
 		return address.get();
 	}
